feat: adds CategoryList component.
notes: this component calls the api before it renders to get the
categories in the redux state, and connects to the redux state to
state to access the data for the categories. It renders a list of
categories and uses the categories data to add links to paths in
which later the categories post will be displayed.